+++ react/components/organisms/profile/profile.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 
 import { baseStyles } from "../../../util";
 
-import { Avatar, Text, Touchable } from "../../atoms";
+import { Avatar, Icon, Text, Touchable } from "../../atoms";
 import { KeyValues } from "../../molecules";
 import { ButtonGroup } from "../button-group";
 
@@ -12,7 +12,13 @@ export class Profile extends Component {
     static get propTypes() {
         return {
             account: PropTypes.object.isRequired,
+            actions: PropTypes.object,
+            backButton: PropTypes.bool,
             editButton: PropTypes.bool,
+            detailsTitle: PropTypes.string,
+            showDescription: PropTypes.bool,
+            showButtons: PropTypes.bool,
+            onBackPress: PropTypes.func,
             onEditPress: PropTypes.func,
             onAvatarPress: PropTypes.func,
             onLogoutTouch: PropTypes.func,
@@ -22,7 +28,15 @@ export class Profile extends Component {
 
     static get defaultProps() {
         return {
+            actions: {},
             editButton: false,
+            backButton: false,
+            detailsTitle: "Account Details",
+            detailsBackgroundColor: "#f6f7f9",
+            showDescription: true,
+            showButtons: true,
+            onPress: () => {},
+            onEditPress: () => {},
             onAvatarPress: () => {},
             styles: styles
         };
@@ -41,51 +55,62 @@ export class Profile extends Component {
         this.scrollRef?.scrollTo({ animated: true, x: 0 });
     };
 
-    _avatarUrl = () => {
+    _avatarUrl() {
         return `${this.props.account?.avatar_url}`;
-    };
+    }
 
-    _details = () => {
+    _details() {
+        console.log(this.props.actions?.email);
         return [
             {
                 key: "E-mail",
-                value: this.props.account.email
+                value: this.props.account.email,
+                ...(this.props.actions?.email || {})
             },
             {
                 key: "Phone",
-                value: this.props.account.meta?.phone_number
+                value: this.props.account.meta?.phone_number,
+                ...(this.props.actions?.phone || {})
             },
             {
                 key: "Company",
-                value: this.props.account.meta?.company
+                value: this.props.account.meta?.company,
+                ...(this.props.actions?.company || {})
             },
             {
                 key: "Company URL",
-                value: this.props.account.meta?.company_url
+                value: this.props.account.meta?.company_url,
+                ...(this.props.actions?.companyUrl || {})
             },
             {
                 key: "Position",
-                value: this.props.account.meta?.position
+                value: this.props.account.meta?.position,
+                ...(this.props.actions?.position || {})
             },
             {
                 key: "Birth Date",
-                value: this.props.account.meta?.birth_date
+                value: this.props.account.meta?.birth_date,
+                ...(this.props.actions?.birthDate || {})
             },
             {
                 key: "Nationality",
-                value: this.props.account.meta?.nationality
+                value: this.props.account.meta?.nationality,
+                ...(this.props.actions?.nationality || {})
             },
             {
                 key: "Roles",
-                value: this.props.account.roles
+                value: this.props.account.roles,
+                ...(this.props.actions?.roles || {})
             },
             {
                 key: "Start Date",
-                value: this.props.account.meta?.start_date
+                value: this.props.account.meta?.start_date,
+                ...(this.props.actions?.startDate || {})
             },
             {
                 key: "Github",
                 value: this.props.account.meta?.github_username,
+                ...(this.props.actions?.githubUsername || {}),
                 valueComponent: (
                     <View style={styles.keyValue}>
                         <Text style={styles.keyPrefixValue}>{"github.com/"}</Text>
@@ -98,6 +123,7 @@ export class Profile extends Component {
             {
                 key: "Twitter",
                 value: this.props.account.meta?.twitter_username,
+                ...(this.props.actions?.twitterUsername || {}),
                 valueComponent: (
                     <View style={styles.keyValue}>
                         <Text style={styles.keyPrefixValue}>{"twitter.com/"}</Text>
@@ -110,6 +136,7 @@ export class Profile extends Component {
             {
                 key: "Linkedin",
                 value: this.props.account.meta?.linkedin_username,
+                ...(this.props.actions?.linkedinUsername || {}),
                 valueComponent: (
                     <View style={styles.keyValue}>
                         <Text style={styles.keyPrefixValue}>{"linkedin.com/in/"}</Text>
@@ -120,9 +147,9 @@ export class Profile extends Component {
                 )
             }
         ].filter(v => Boolean(v));
-    };
+    }
 
-    _buttons = () => {
+    _buttons() {
         const buttons = [];
         if (this.props.account.meta?.phone_number) {
             buttons.push({
@@ -140,9 +167,19 @@ export class Profile extends Component {
             }
         });
         return buttons;
-    };
+    }
 
-    _buttonGroupStyle = () => {
+    _showDescription() {
+        return this.props.account && this.props.account.description && this.props.showDescription;
+    }
+
+    async onShareContactPress() {
+        await Share.share({
+            message: this.props.account.meta?.phone_number
+        });
+    }
+
+    _buttonGroupStyle() {
         return {
             borderBottomLeftRadius: 0,
             borderBottomRightRadius: 0,
@@ -150,17 +187,11 @@ export class Profile extends Component {
             borderTopRightRadius: 0,
             paddingHorizontal: 15
         };
-    };
+    }
 
-    onShareContactPress = async () => {
-        await Share.share({
-            message: this.props.account.meta?.phone_number
-        });
-    };
-
-    _renderHeader = () => {
+    _renderHeader() {
         return (
-            <>
+            <View style={styles.header}>
                 <Avatar
                     image={{
                         uri: `${this.props.account?.avatar_url}`
@@ -169,51 +200,70 @@ export class Profile extends Component {
                     onPress={this.props.onAvatarPress}
                 />
                 <Text style={styles.username}>{this.props.account?.meta?.name || ""}</Text>
-                {this.props.account && this.props.account.description ? (
+                {this._showDescription() ? (
                     <Text style={styles.description}>
                         {this.props.account?.description.trim() || ""}
                     </Text>
                 ) : null}
-            </>
+            </View>
         );
-    };
+    }
 
-    _renderDetails = () => {
+    _renderDetails() {
         return (
             <>
                 <KeyValues items={this._details()} showUnset={false} />
-                <View style={styles.buttons}>
-                    <ButtonGroup
-                        buttonStyle={this._buttonGroupStyle()}
-                        items={this._buttons()}
-                        orientation={"vertical"}
-                        variant={"flat"}
-                        align={"left"}
-                        toggle={false}
-                    />
-                </View>
+                {this.props.showButtons && (
+                    <View style={styles.buttons}>
+                        <ButtonGroup
+                            buttonStyle={this._buttonGroupStyle()}
+                            items={this._buttons()}
+                            orientation={"vertical"}
+                            variant={"flat"}
+                            align={"left"}
+                            toggle={false}
+                        />
+                    </View>
+                )}
             </>
         );
-    };
+    }
 
     render() {
         return (
-            <>
-                <View style={styles.header}>
+            <View style={styles.profile}>
+                <View style={styles.headerContainer}>
+                    <View style={styles.headerButtons}>
+                        {this.props.backButton && (
+                            <Touchable
+                                style={styles.backButton}
+                                hitSlop={{ top: 20, left: 20, right: 20, bottom: 20 }}
+                                onPress={() => this.props.onBackPress()}
+                            >
+                                <Icon
+                                    style={styles.goBackButtonIcon}
+                                    color="#1d2631"
+                                    icon="arrow-left"
+                                    width={24}
+                                    height={24}
+                                    strokeWidth={2}
+                                />
+                            </Touchable>
+                        )}
+                        {this.props.editButton && (
+                            <Touchable
+                                style={styles.editButton}
+                                hitSlop={{ top: 20, left: 20, right: 20, bottom: 20 }}
+                                onPress={this.props.onEditPress}
+                            >
+                                <Text style={styles.editButtonText}>Edit</Text>
+                            </Touchable>
+                        )}
+                    </View>
                     {this._renderHeader()}
-
-                    {this.props.editButton && (
-                        <Touchable
-                            hitSlop={{ top: 20, left: 20, right: 20, bottom: 20 }}
-                            onPress={this.props.onEditPress}
-                            style={styles.editButton}
-                        >
-                            <Text style={styles.editButtonText}>Edit</Text>
-                        </Touchable>
-                    )}
                 </View>
                 <View style={styles.detailsTitle}>
-                    <Text style={styles.detailsTitleText}>Account Details</Text>
+                    <Text style={styles.detailsTitleText}>{this.props.detailsTitle}</Text>
                 </View>
                 <ScrollView
                     style={styles.details}
@@ -223,20 +273,43 @@ export class Profile extends Component {
                 >
                     {this._renderDetails()}
                 </ScrollView>
-            </>
+            </View>
         );
     }
 }
 
 const styles = StyleSheet.create({
+    profile: {
+        flex: 1,
+        backgroundColor: "#ffffff"
+    },
+    headerContainer: {
+        alignItems: "center"
+    },
     header: {
         alignItems: "center",
-        paddingVertical: 20
+        paddingVertical: 15
+    },
+    headerButtons: {
+        width: "100%",
+        backgroundColor: "#f6f7f9"
+    },
+    backButton: {
+        position: "relative",
+        height: 50,
+        width: 66
+    },
+    goBackButtonIcon: {
+        position: "absolute",
+        left: 21,
+        top: 10
     },
     editButton: {
-        position: "absolute",
+        alignSelf: "flex-end",
+        position: "relative",
+        paddingVertical: 14,
         right: 18,
-        top: 10
+        top: 0
     },
     editButtonText: {
         fontSize: 16,